refactor(benchmarking-cli): scope --sanity-weight-check to PalletCmd
Moving SanityWeightCheck out of sc_cli::SharedParams and into PalletCmd,
where it actually has effect. Previously the flag appeared in the
--help output of every Substrate node binary (run, key, chain-info,
purge-chain, etc.) and was silently ignored everywhere except `benchmark
pallet`. The enum lives in frame-benchmarking-cli now since that is the
only consumer; sc-cli remains free of benchmarking-specific knobs.

Refs #152
